api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic_settings
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
requests
werkzeug
psycopg2-binary
passlib[argon2]
